Modularizacija cijelog frontenda prema MVVM strukturi
