+++ frontend/src/viewModels/useLogin.js
@@ -15,7 +15,7 @@ export default function Login() {
     }
 
     try {
-      const response = await axios.post("http://localhost:3000/login", {
+      const response = await axios.post("http://localhost:3000/api/login", {
         username,
         password,
       });
